Move existing-credit-card to checkout payment-methods directory
